+++ app/src/main/java/com/example/data/MonthlyRecord.kt
@@ -6,6 +6,6 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "monthly_records")
 data class MonthlyRecord(
     @PrimaryKey val monthYear: String, // e.g. "July 2026", "August 2026"
-    val monthlyIncome: Double? = null,
+    val monthlyIncome: Long? = null, // in paise
     val note: String? = null
 )
